pyproject.toml
[project]
name = "rl-capstone"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "pillow",
    "opencv-python",
    "pygame",
    "gymnasium",
    "tensorflow-macos; platform_system == 'Darwin' and platform_machine == 'arm64'",
    "tensorflow-metal; platform_system == 'Darwin' and platform_machine == 'arm64'",
    "keras",
    "matplotlib",
    "tqdm",
    "torch",                # PyTorch (CPU/MPS build will auto-resolve on macOS)
    "stable-baselines3",
]

[project.optional-dependencies]
dev = ["ruff"]

[tool.ruff]
line-length = 120

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["rl_capstone"]
